add trait for pagination and delete AbstractPaginatorController
